requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.0.2
django-import-export==2.5.0
et-xmlfile==1.0.1
future==0.18.2
html5lib==1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.1.0
psycopg2==2.8.6
PyPDF2==1.26.0
python-bidi==0.4.2
python-decouple==3.3
pytz==2020.4
PyYAML==5.4.1
reportlab==3.5.59
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
